requirements.txt
# Core AI
openai>=1.40.0
pydantic>=2.0.0
numpy>=1.24.0

# HTTP Client (async)
httpx>=0.27.0

# TUI
rich>=13.0.0
textual>=0.80.0

# Environment
python-dotenv>=1.0.0

# Vector math (cosine similarity, optional clustering)
scikit-learn>=1.3.0

# Hebrew RTL Terminal Support
python-bidi>=0.4.2

# Human-readable relative timestamps
timeago>=1.0.0